version=12.1.51_dev
# new feature: isosurface slab TRANLSUCENT 0.6 red x=3
#  -- preliminary translucent slabbing



Former-commit-id: 8ba921a8feb51027d2a0170d8d11be5d3a4839d5

+++ Jmol/src/org/jmol/jvxl/data/JvxlData.java
@@ -153,7 +153,7 @@ public class JvxlData {
   public int nVertexColors;
   public String color;
   public String meshColor;
-  public int translucency;
+  public float translucency;
   public String colorScheme;
   public String rendering;
   public int slabValue = Integer.MIN_VALUE;
